Additionally check wanted episodes against a backup title.
This primarily affects the BTN provider, as often times there are uploads of torrents with invalid/non-scene release names.
This change makes the generic provider double check if the episode is wanted against the backup title.
The backup title is the title that's created by concating all the json info from the BTN api.